Address issues with CMAKE_BUILD_TYPE
The CMAKE_BUILD_TYPE definition in the top-level CMakeLists.txt
file wasn't working as expected. It turns out that it has to be
defined _before_ the project() command or it has no effect.

- Moved the definition of CMAKE_BUILD_TYPE before the project()
  command.
- Defined a new KLOCWORK command-line option, which defaults to
  OFF.
- Modified CMAKE_BUILD_TYPE to default to "Debug" if the KLOCWORK
  option is enabled or "RelWithDebInfo" if is is disabled.

When building in Debug mode, CMake suppresses the definition of
the NDEBUG flag, thereby enabling any assert() statements in the
code. We do this so Klocwork takes variable references in the
assertions into account when doing static analysis.

Signed-off-by: Derek G Foster <derek.foster@intel.com>