Fix SDE dependent library lookup (#48)
A few of the SDE libraries on which we are directly dependent do not have
SONAME field populated. Because of this, when infrap4d links to those
libraries, it links with absolute path. This causes an issue when we build
networking-recipe on one machine and copy the binaries/libraries to other
machine.

To avoid this, in es2k cmake file we set IMPORTED_NO_SONAME property to
specifically tell the linker that the imported shared library file
has no soname field.

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>