Provide new options for dependencies cross compilation script (#50)
With this PR we are introducing below two options for make-cross-deps.sh
script

1) no-download: By default we download the source code for all dependent
modules and build them. But if user has dependent code already downloaded,
then user can provide this option to the script which will not download the
source code on the fly.
  ex: ./make-cross-deps.sh --no-download

2) force: gRPC Patch is applied by the CMake by default. If user has
already applied the patch, and tries to apply for the second time,
cmake halts the execution and waits for user input as [y/n] to apply the
patch. If user wants to forcefully apply the patch without halting the
cmake build, then user can provide this option.
  ex: ./make-cross-deps.sh --force

Also add missing support for the long-form `--jobs` parameter.

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>