Fix bug in gRPC v1.54.2 cmake build script
Cross-compilation of gRPC v1.54.2 fails with the error message

    INSTALL(EXPORT) given unknown export "gRPCPluginTargets"

We suppress the gRPC plugins when cross-compiling for ACC because
we use the HOST compiler to do the build, not the target compiler.
This means that there are no gRPC Plugin Targets to be exported.

We solve the problem by conditionalizing the INSTALL(EXPORT) command
with the same variable used to conditionalize generation of the
plugins.

Signed-off-by: Derek G Foster <derek.foster@intel.com>