Address review comments and fix dpdk build
Changes include segregating ES2K and DPDK code in ovs-p4rt
Quick fix in p4rt-ctl to look for only group_id for action selectors
which is 'comma' seperated.
This check was causing issue when we try to program lag_group_id
which is also having group_id as key word.

Signed-off-by: Sandeep N <sandeep.nagapattinam@intel.com>