Overhauled OVSP4RT logging to fix design issue
Problem statement: the ES2K implementation of ovsp4rt_config_fdb_entry()
logs messages that include the hard-coded names of the underlying P4
tables. There are several problems with this:

1. It's not target/program-agnostic: it assumes that all P4 programs
   for all targets use the same table names for the same purposes,
   and will continue to do so in the future.

2. It bakes the table names into the code, which makes the code less
   maintainable.

3. It doesn't work in cases where the underlying code chooses between
   P4 tables based on information in the parameter object (e.g., ipv4
   vs. ipv6).

4. The mid-level functions implement an abstraction model that
   explicitly hides the table names from the top-level interface.
   Encoding the underlying table names in top-level messages violates
   the abstraction.

This CL moves logging to the mid-level functions. It also implements
an interface that allows the third-level functions to return a
handle that a mid-level function can use to obtain the name of the
underlying table.

Signed-off-by: Derek Foster <derek.foster@intel.com>