Limited integration of ovs-p4rt and Tracker class
Demonstrates how the Tracker class could be integrated with
minimal changes to ovsp4rt.cc, as might be done to collect test
data using the existing code.

The public API passes a Context object containing a reference to
the Tracker object, rather than Tracker object itself, because we
want the code that uses the Tracker object to be as close to its
final form as circumstances permit.

The Context object includes references to a dummy Envoy object
and the P4Info object, which the internal function does not use
(another case of wanting the code to be close to its final form).

The next step in the evolution of ovsp4rt.cc would be modify the
internal functions to use the p4info object in place of the
discrete parameter, and the envoy object in place of the discrete
session parameter.

Signed-off-by: Derek Foster <derek.foster@intel.com>