Revise src_port verification
The src_port field contains an arbitrary value that has nothing
to do with what was specified in the tunnel_info structure.
This is a workaround for a long-standing bug in the Linux
Networking P4 program.

Documented this fact in the test program and placed the check
under control of a --check_src_port command-line flag that
defaults to False.

Signed-off-by: Derek Foster <derek.foster@intel.com>